environmentd: move api/sql backing impl into http mod
